Major change: Making the releases/6.0.0-alpha1 as core folder and editing index and package files
