+++ README.md
@@ -1,10 +1,10 @@
 # Kibana Sankey Diagram Plugin
 
-This is a sankey diagram visType plugin for Kibana 4.3+.
+This is a sankey diagram visType plugin for Kibana 6.0.0-alpha1.
 
 This plugin was developped from <https://github.com/elastic/kibana/pull/4832>.
 
-![](https://cloud.githubusercontent.com/assets/1219655/9702343/081607e6-548b-11e5-81cb-4523c8c9225d.png)
+![](https://github.com/JuanCarniglia/kbn_sankey_vis/blob/master/releases/6.0.0-alpha1/Capture1.PNG)
 
 # Install
