--- bower.json
@@ -1,7 +1,0 @@
-{
-  "name": "kbn_sankey_vis",
-  "version": "0.1.0",
-  "dependencies": {
-    "d3-plugins-sankey": ">=1.1.0"
-  }
-}
