+++ index.js
@@ -1,12 +1,11 @@
-module.exports = function (kibana) {
-  return new kibana.Plugin({
-    name: 'kbn_sankey_vis',
-    require: ['kibana', 'elasticsearch'],
-    uiExports: {
-      visTypes: [
-        'plugins/kbn_sankey_vis/kbn_sankey_vis'
-      ]
-    }
-  });
+export default function (kibana) {
+
+        return new kibana.Plugin({
+                uiExports: {
+                        visTypes: [
+                                'plugins/kbn_sankey_vis/kbn_sankey_vis'
+                ]
+        }
+        });
 };
