+++ package.json
@@ -1,37 +1,28 @@
 {
   "name": "kbn_sankey_vis",
-  "version": "0.1.0",
+  "version": "kibana",
   "authors": [
     "Chenryn <rao.chenlin@gmail.com>"
   ],
-  "description": "sankey diagram visualization plugin",
-  "main": "index.js",
+  "description": "Sankey Diagram Visualization Plugin",
   "license": "Apache-2.0",
-  "homepage": "http://kibana.logstash.es/",
+  "homepage": "https://github.com/WolfangAukang/kbn_sankey_vis",
+  "repository": "https://github.com/WolfangAukang/kbn_sankey_vis.git",
   "scripts": {
     "test": "gulp test",
     "test:coverage": "gulp coverage",
     "start": "gulp dev",
-    "postinstall": "bower install",
     "precommit": "gulp lint",
     "build": "gulp build"
   },
   "dependencies": {
-    "d3-plugins-sankey": ">=1.1.0"
-  },
-  "devDependencies": {
+    "d3-plugins-sankey": ">=1.1.0",
     "babel-eslint": "^4.1.8",
-    "babel-preset-es2015": "^6.5.0",
-    "babel-register": "^6.5.2",
     "bluebird": "^3.3.1",
-    "bower": "^1.6.9",
-    "del": "^2.2.0",
     "eslint": "^2.1.0",
     "eslint-plugin-mocha": "^2.0.0",
-    "gulp": "^3.9.1",
+    "gulp": "3.9.0",
     "gulp-eslint": "^2.0.0",
-    "gulp-gzip": "^1.2.0",
-    "gulp-tar": "^1.8.0",
     "gulp-util": "^3.0.7",
     "gulp-zip": "3.1.0",
     "husky": "^0.10.2",
