--- releases/6.0.0-alpha1/.jshintrc
@@ -1,3 +1,0 @@
-{
-  "esversion" : 6
-}
