--- releases/6.0.0-alpha1/index.js
@@ -1,11 +1,0 @@
-export default function (kibana) {
-  return new kibana.Plugin({
-    name: 'kbn_sankey_vis',
-    require: ['kibana', 'elasticsearch'],
-    uiExports: {
-      visTypes: [
-        'plugins/kbn_sankey_vis/kbn_sankey_vis'
-      ]
-    }
-  });
-}
