--- releases/6.0.0-alpha1/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "kbn_sankey_vis",
-  "version": "kibana",
-  "authors": [
-    "Chenryn <rao.chenlin@gmail.com>"
-  ],
-  "description": "Sankey Diagram Visualization Plugin",
-  "main": "index.js",
-  "license": "Apache-2.0",
-  "homepage": "https://github.com/JuanCarniglia/kbn_sankey_vis",
-  "repository": "https://github.com/JuanCarniglia/kbn_sankey_vis.git",
-  "scripts": {
-    "test": "gulp test",
-    "test:coverage": "gulp coverage",
-    "start": "gulp dev",
-    "precommit": "gulp lint",
-    "build": "gulp build"
-  },
-  "dependencies": {
-    "d3-plugins-sankey": ">=1.1.0"
-  },
-  "devDependencies": {
-    "babel-eslint": "^4.1.8",
-    "bluebird": "^3.3.1",
-    "eslint": "^2.1.0",
-    "eslint-plugin-mocha": "^2.0.0",
-    "gulp": "^3.9.1",
-    "gulp-eslint": "^2.0.0",
-    "gulp-util": "^3.0.7",
-    "gulp-zip": "3.1.0",
-    "husky": "^0.10.2",
-    "lodash": "^4.4.0",
-    "minimist": "1.2.0",
-    "mkdirp": "^0.5.1",
-    "rimraf": "^2.4.4",
-    "rsync": "^0.4.0",
-    "d3-plugins-sankey": ">=1.1.0"
-  }
-}
