--- releases/6.0.0-alpha1/public/kbn_sankey_vis.html
@@ -1,3 +1,0 @@
-<div ng-controller="KbnSankeyVisController" class="metric-vis">
-  <div class="metric-container" ng-model="metrics"></div>
-</div>
